Used delegates and c# events for each instance to subscribe to one event and listen to the changes there
In the playerconfigurationmanager class, I made a delegate and event that have been subscribed by each instances of the playersetupmenu class. Each time a new player joins, a new instance of the playersetupmenu is created. So, all I did was pass in the value of the selected button to the event listener, (which is the same for every instance) and when the event is called, it notifies every instance listening to it and they all act accordingly to it.