Solved the multiple damage bug
oh , i think i've solved it. I thought the instantiation was being done twice or thrice according to the number of players in the room.
But the real problem was the function triggering multiple times not the bullet being instantiated multiple times.

so i just simply added an additional "photonView.IsMine" check in the trigger function